requirements.txt
pySerial
pycryptodome
paho-mqtt
python-dateutil
